pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "docling>=2.39.0",
    "easyocr>=1.7.2",
    "google-genai>=1.24.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "onnxruntime-gpu>=1.22.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "rapidocr-onnxruntime>=1.4.4",
    "sentence-transformers>=5.0.0",
    "sqlmodel>=0.0.24",
    "streamlit>=1.46.1",
]
